requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8
Pillow>=9.0
